+++ packages/ozone/src/api/moderation/getRecord.ts
@@ -7,11 +7,12 @@ import { AtUri } from '@atproto/syntax'
 export default function (server: Server, ctx: AppContext) {
   server.tools.ozone.moderation.getRecord({
     auth: ctx.authVerifier.modOrAdminToken,
-    handler: async ({ params, auth }) => {
+    handler: async ({ params, auth, req }) => {
       const db = ctx.db
+      const labelers = ctx.reqLabelers(req)
 
       const [record, accountInfo] = await Promise.all([
-        ctx.modService(db).views.recordDetail(params),
+        ctx.modService(db).views.recordDetail(params, labelers),
         getPdsAccountInfo(ctx, new AtUri(params.uri).hostname),
       ])